refactor(core): promote the deferred imports the split made safe
Eight function-level imports written to dodge a cycle through common.py no
longer sit on a cycle, because the symbol they reach for lives in a module
below them.
