+++ geoparquet_io/core/write_strategies/arrow_streaming.py
@@ -18,6 +18,7 @@ from typing import TYPE_CHECKING
 import pyarrow as pa
 import pyarrow.parquet as pq
 
+from geoparquet_io.core.compression import validate_compression_settings
 from geoparquet_io.core.duckdb_utils import quote_identifier
 from geoparquet_io.core.geoarrow_encoding import (
     WKB_EXTENSION_NAMES,
@@ -267,7 +268,6 @@ class ArrowStreamingStrategy(BaseWriteStrategy):
         extra_kv_metadata: dict[str, str] | None = None,
     ) -> None:
         """Write query results to GeoParquet using streaming RecordBatch approach."""
-        from geoparquet_io.core.common import validate_compression_settings
         from geoparquet_io.core.geo_metadata import GEOPARQUET_VERSIONS
 
         configure_verbose(verbose)
@@ -667,7 +667,6 @@ class ArrowStreamingStrategy(BaseWriteStrategy):
         from geoparquet_io.core.common import (
             _compute_geometry_types,
             _detect_version_from_table,
-            validate_compression_settings,
         )
         from geoparquet_io.core.crs_utils import apply_output_crs
         from geoparquet_io.core.geo_metadata import GEOPARQUET_VERSIONS, create_geo_metadata
